# specs/001-media-influence-detection/contracts/api.yaml
openapi: 3.1.0
info:
  title: MITDS API
  description: Media Influence Topology & Detection System REST API
  version: 1.0.0
  contact:
    name: MITDS Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.mitds.example/v1
    description: Production
...
paths:
...
  /entities:
    get:
      summary: Search entities
      tags: [Entities]
      operationId: searchEntities
      parameters:
        - name: q
          in: query
          description: Search query (name, alias, identifier)
          schema:
            type: string
        - name: type
          in: query
          description: Filter by entity type
          schema:
            $ref: '#/components/schemas/EntityType'
        - name: jurisdiction
          in: query
          description: Filter by jurisdiction (for organizations)
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitySearchResponse'
  /entities/{id}:
    get:
      summary: Get entity by ID
      tags: [Entities]
      operationId: getEntity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Entity details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '404':
          description: Entity not found
  /entities/{id}/relationships:
    get:
      summary: Get entity relationships
      tags: [Entities, Relationships]
      operationId: getEntityRelationships
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rel_type
          in: query
          description: Filter by relationship type
          schema:
            $ref: '#/components/schemas/RelationshipType'
        - name: direction
          in: query
          description: Relationship direction from entity
          schema:
            type: string
            enum: [inbound, outbound, both]
            default: both
        - name: as_of
          in: query
          description: Point-in-time query (ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Relationships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipListResponse'
  /entities/{id}/evidence:
    get:
      summary: Get evidence for entity
      tags: [Entities]
      operationId: getEntityEvidence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Evidence list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceListResponse'

  # ============== RELATIONSHIPS ==============
...
components:
  schemas:
...
    EntityType:
      type: string
      enum: [PERSON, ORGANIZATION, OUTLET, DOMAIN, PLATFORM_ACCOUNT, SPONSOR, VENDOR]
    RelationshipType:
      type: string
      enum: [FUNDED_BY, DIRECTOR_OF, EMPLOYED_BY, SPONSORED_BY, OWNS, CITED, AMPLIFIED, SHARED_INFRA]
...
    Entity:
      type: object
      required: [id, entity_type, name, created_at, confidence]
      properties:
        id:
          type: string
          format: uuid
        entity_type:
          $ref: '#/components/schemas/EntityType'
        name:
          type: string
        aliases:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        confidence:
          type: number
          minimum: 0
          maximum: 1
        # Type-specific properties included dynamically
        properties:
          type: object
          additionalProperties: true
    EntitySearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ============== RELATIONSHIPS ==============
    Relationship:
      type: object
      required: [id, rel_type, source_entity, target_entity, confidence]
      properties:
        id:
          type: string
          format: uuid
        rel_type:
          $ref: '#/components/schemas/RelationshipType'
        source_entity:
          $ref: '#/components/schemas/EntitySummary'
        target_entity:
          $ref: '#/components/schemas/EntitySummary'
        valid_from:
          type: string
          format: date-time
        valid_to:
          type: string
          format: date-time
        confidence:
          type: number
          minimum: 0
          maximum: 1
        properties:
          type: object
          additionalProperties: true
        evidence_count:
          type: integer
    EntitySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entity_type:
          $ref: '#/components/schemas/EntityType'
        name:
          type: string
    RelationshipListResponse:
      type: object
      properties:
        relationships:
          type: array
          items:
            $ref: '#/components/schemas/Relationship'
        total:
          type: integer
...
    EvidenceRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        evidence_type:
          type: string
        source_url:
          type: string
          format: uri
        archive_url:
          type: string
          format: uri
        retrieved_at:
          type: string
          format: date-time
    EvidenceListResponse:
      type: object
      properties:
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceRef'

    # ============== REPORTS ==============
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []